Add Call Checklist feature to Action Popup and enhance UI components
