+++ hora-new-admin-dashboard/src/app/component/Actionpopup.css
@@ -14,8 +14,8 @@
     background-color: #fff;
     padding: 10px;
     border-radius: 12px;
-    max-width: 900px;
-    max-height: 500px;
+    max-width: 1000px;
+    max-height: 600px;
     width: 90%;
     box-shadow: 0px 8px 30px rgba(0, 0, 0, 0.2);
     position: relative;
@@ -46,7 +46,7 @@
   
   .order-grid {
     display: grid;
-    grid-template-columns: 2fr 1fr;
+    grid-template-columns: 2fr 1.5fr;
     gap: 20px;
   }
   
@@ -110,7 +110,6 @@
     color: white;
     border-radius: 8px;
     box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.1);
-    text-align: center;
   }
   
   .order-summary-box h3 {
@@ -141,4 +140,12 @@
   }
 .margin-top{
   margin-top: 20px;
+}
+.callChecklist-container{
+  max-height: 300px;
+  overflow-y: auto;
+
+}
+.text-center{
+text-align: center;
 }
